+++ packages/e2e-tests/src/enums.ts
@@ -23,6 +23,7 @@ export enum SettingsModalTabName {
   NOTIFICATIONS = 'notifications',
   QR_CODE = 'qr-code',
   LEAVE_COMMUNITY = 'leave-community',
+  DEBUG = 'debug',
 }
 
 export enum PhotoExt {