+++ packages/state-manager/src/sagas/appConnection/connection.selectors.ts
@@ -2,7 +2,7 @@ import { createSelector } from 'reselect'
 import { StoreKeys } from '../store.keys'
 import { type CreatedSelectors, type StoreState } from '../store.types'
 import { peersStatsAdapter } from './connection.adapter'
-import { isCurrentCommunityInitialized } from '../network/network.selectors'
+import { connectedPeers, isCurrentCommunityInitialized } from '../network/network.selectors'
 import { composeInvitationShareUrl, createLibp2pAddress, filterAndSortPeers, p2pAddressesToPairs } from '@quiet/common'
 import { areMessagesLoaded, areChannelsLoaded } from '../publicChannels/publicChannels.selectors'
 import { identitySelectors } from '../identity/identity.selectors'
@@ -25,6 +25,8 @@ export const connectionProcess = createSelector(connectionSlice, reducerState =>
 
 export const socketIOSecret = createSelector(connectionSlice, reducerState => reducerState.socketIOSecret)
 
+export const p2pEnabled = createSelector(connectionSlice, reducerState => reducerState.p2pEnabled)
+
 export const peerStats = createSelector(connectionSlice, reducerState => {
   let stats: NetworkStats[]
   if (reducerState.peersStats === undefined) {
@@ -39,7 +41,8 @@ export const peerList = createSelector(
   userProfileSelectors.userProfiles,
   identitySelectors.currentPeerAddress,
   peerStats,
-  (userProfiles, localPeerAddress, stats) => {
+  connectedPeers,
+  (userProfiles, localPeerAddress, stats, connectedPeers) => {
     let arr: string[] = []
     if (userProfiles) {
       const profiles = Object.values(userProfiles)
@@ -52,7 +55,7 @@ export const peerList = createSelector(
         })
         .filter((address): address is string => address !== null && address !== undefined)
     }
-    const filteredAndSortedPeers = filterAndSortPeers(arr, stats, localPeerAddress)
+    const filteredAndSortedPeers = filterAndSortPeers(arr, stats, localPeerAddress, true, connectedPeers)
     return filteredAndSortedPeers
   }
 )
@@ -140,4 +143,5 @@ export const connectionSelectors = {
   socketIOSecret,
   isJoiningCompleted,
   peerStats,
+  p2pEnabled,
 }