+++ packages/state-manager/src/sagas/appConnection/connection.slice.ts
@@ -24,6 +24,7 @@ export class ConnectionState {
     text: ConnectionProcessInfo.CONNECTION_STARTED,
   }
   public longLivedInvite: InviteResult | undefined = undefined
+  public p2pEnabled: boolean = true
 }
 
 export const connectionSlice = createSlice({
@@ -37,6 +38,7 @@ export const connectionSlice = createSlice({
       const _peerStats = state.peersStats || peersStatsAdapter.getInitialState()
       peersStatsAdapter.upsertOne(_peerStats, {
         peerId: action.payload.peer,
+        address: action.payload.address,
         lastSeen: action.payload.lastSeen,
         connectionTime: 0,
       })
@@ -46,6 +48,7 @@ export const connectionSlice = createSlice({
       const _peerStats = state.peersStats || peersStatsAdapter.getInitialState()
       peersStatsAdapter.upsertOne(_peerStats, {
         peerId: action.payload.peer,
+        address: action.payload.address,
         lastSeen: action.payload.lastSeen,
         connectionTime: prev + action.payload.connectionDuration,
       })
@@ -84,6 +87,10 @@ export const connectionSlice = createSlice({
       }
     },
     createInvite: (state, _action: PayloadAction<any>) => state,
+    toggleP2P: state => state,
+    setP2PEnabled: (state, action: PayloadAction<boolean>) => {
+      state.p2pEnabled = action.payload
+    },
   },
 })